Outsourced auth logic to store/auth.js
