+++ src/store/auth.js
@@ -1,7 +1,35 @@
-import React from "react";
+import React, { useState } from "react";
 
 const AuthContext = React.createContext({
-    isLoggedIn: false,
+  isLoggedIn: false,
+  onLogin: () => {},
+  onLogout: () => {},
 });
+
+export const AuthContextProvider = (props) => {
+  const [isLoggedIn, setIsLoggedIn] = useState(false);
+
+  const loginHandler = () => {
+      setIsLoggedIn(true);
+      localStorage.setItem("login_key", 1);
+  };
+
+  const logoutHandler = () => {
+    setIsLoggedIn(false);
+    localStorage.setItem("login_key", 0);
+  };
+
+  return (
+    <AuthContext.Provider
+      value={{
+        isLoggedIn: isLoggedIn,
+        onLogin: loginHandler,
+        onLogout: logoutHandler,
+      }}
+    >
+      {props.children}
+    </AuthContext.Provider>
+  );
+};
 
 export default AuthContext;
